+++ service/telemetry.go
@@ -5,7 +5,6 @@ package service // import "go.opentelemetry.io/collector/service"
 
 import (
 	"context"
-	"errors"
 	"net"
 	"net/http"
 	"strconv"
@@ -13,41 +12,26 @@ import (
 	ocmetric "go.opencensus.io/metric"
 	"go.opencensus.io/metric/metricproducer"
 	"go.opentelemetry.io/contrib/config"
-	"go.opentelemetry.io/contrib/propagators/b3"
-	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/metric"
 	noopmetric "go.opentelemetry.io/otel/metric/noop"
-	"go.opentelemetry.io/otel/propagation"
 	sdkmetric "go.opentelemetry.io/otel/sdk/metric"
 	"go.opentelemetry.io/otel/sdk/resource"
-	"go.opentelemetry.io/otel/trace"
-	nooptrace "go.opentelemetry.io/otel/trace/noop"
 	"go.uber.org/multierr"
 	"go.uber.org/zap"
 
 	"go.opentelemetry.io/collector/config/configtelemetry"
 	"go.opentelemetry.io/collector/service/internal/proctelemetry"
-	"go.opentelemetry.io/collector/service/internal/servicetelemetry"
 	"go.opentelemetry.io/collector/service/telemetry"
 )
 
 const (
 	zapKeyTelemetryAddress = "address"
 	zapKeyTelemetryLevel   = "level"
-
-	// supported trace propagators
-	traceContextPropagator = "tracecontext"
-	b3Propagator           = "b3"
-)
-
-var (
-	errUnsupportedPropagator = errors.New("unsupported trace propagator")
 )
 
 type telemetryInitializer struct {
 	ocRegistry *ocmetric.Registry
 	mp         metric.MeterProvider
-	tp         trace.TracerProvider
 	servers    []*http.Server
 
 	disableHighCardinality bool
@@ -57,15 +41,14 @@ type telemetryInitializer struct {
 func newColTelemetry(disableHighCardinality bool, extendedConfig bool) *telemetryInitializer {
 	return &telemetryInitializer{
 		mp:                     noopmetric.NewMeterProvider(),
-		tp:                     nooptrace.NewTracerProvider(),
 		disableHighCardinality: disableHighCardinality,
 		extendedConfig:         extendedConfig,
 	}
 }
 
-func (tel *telemetryInitializer) init(res *resource.Resource, settings servicetelemetry.TelemetrySettings, cfg telemetry.Config, asyncErrorChannel chan error) error {
+func (tel *telemetryInitializer) init(res *resource.Resource, logger *zap.Logger, cfg telemetry.Config, asyncErrorChannel chan error) error {
 	if cfg.Metrics.Level == configtelemetry.LevelNone || (cfg.Metrics.Address == "" && len(cfg.Metrics.Readers) == 0) {
-		settings.Logger.Info(
+		logger.Info(
 			"Skipping telemetry setup.",
 			zap.String(zapKeyTelemetryAddress, cfg.Metrics.Address),
 			zap.String(zapKeyTelemetryLevel, cfg.Metrics.Level.String()),
@@ -73,28 +56,9 @@ func (tel *telemetryInitializer) init(res *resource.Resource, settings servicete
 		return nil
 	}
 
-	settings.Logger.Info("Setting up own telemetry...")
-	configuredSDK, err := config.NewSDK(
-		config.WithOpenTelemetryConfiguration(config.OpenTelemetryConfiguration{
-			TracerProvider: &config.TracerProvider{
-				Processors: cfg.Traces.Processors,
-			},
-		}),
-	)
+	logger.Info("Setting up own telemetry...")
 
-	if err != nil {
-		return err
-	}
-
-	tel.tp = configuredSDK.TracerProvider()
-
-	if tp, err := textMapPropagatorFromConfig(cfg.Traces.Propagators); err == nil {
-		otel.SetTextMapPropagator(tp)
-	} else {
-		return err
-	}
-
-	return tel.initMetrics(res, settings.Logger, cfg, asyncErrorChannel)
+	return tel.initMetrics(res, logger, cfg, asyncErrorChannel)
 }
 
 func (tel *telemetryInitializer) initMetrics(res *resource.Resource, logger *zap.Logger, cfg telemetry.Config, asyncErrorChannel chan error) error {
@@ -165,19 +129,4 @@ func (tel *telemetryInitializer) shutdown() error {
 		}
 	}
 	return errs
-}
-
-func textMapPropagatorFromConfig(props []string) (propagation.TextMapPropagator, error) {
-	var textMapPropagators []propagation.TextMapPropagator
-	for _, prop := range props {
-		switch prop {
-		case traceContextPropagator:
-			textMapPropagators = append(textMapPropagators, propagation.TraceContext{})
-		case b3Propagator:
-			textMapPropagators = append(textMapPropagators, b3.New())
-		default:
-			return nil, errUnsupportedPropagator
-		}
-	}
-	return propagation.NewCompositeTextMapPropagator(textMapPropagators...), nil
 }
